magikarp: drop the custom keyboard config
Magikarp has a custom keyboard config, but this is only used when
vivaldi keyboard support is enabled, which depends on
PLATFORM_EC_KEYBOARD_PROTOCOL_8042, which is not enabled on any corsola
project, so this is dead code. Drop it.

BUG=b:274443029
TEST=zmake compare-builds magikarp

Change-Id: Ic94cc98ece44f2c0b4c9bedc3634a8e7c6c12a18
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5171787
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>