pdc: Add driver function for getting bus info
The EC_CHIP_LOCATE host command needs to be able to obtain the I2C bus info
of the PDC (I2C port number and address) so that the AP can talk to it in
passthrough mode. Expand the PDC driver API with a get_bus_info method that
supplies this data.

BUG=b:304301475
TEST=zmake build brya_pdc
LOW_COVERAGE_REASON=Coverage to be handled in b/317065172

Change-Id: I7a2497c5c836a623104f9203f33ea010e61ac2e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5202123
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Sam Hurst <shurst@google.com>