when switching EPR ports directly (#941)
* when switching EPR ports directly

always make sure the charger has exited bypass mode
to avoid transients being passed to the system

Signed-off-by: Kieran Levin <ktl@frame.work>

* handle ACOK threshold

Set ACOK threshold when transitioning to NVDC to discharge input power.
Make sure charger task is not woken from inside charge_manager task
board hook.

If this happens the state will be out of sync between charge manager
and charger.
And charger will try to update with the old values

Signed-off-by: Kieran Levin <ktl@frame.work>

* force disable the charger bypass when switching

Different parts of this run on different threads, so just force bypass
off for two runs of the charger updating.

Signed-off-by: Kieran Levin <ktl@frame.work>

* fix azalea build

limit power to 100W when charger not in bypass mode

Signed-off-by: Kieran Levin <ktl@frame.work>

---------

Signed-off-by: Kieran Levin <ktl@frame.work>