zephyr: fpmcu: do not log GPR during panic
The general purpose registers may contain secret when an exception
occurs. Do not log these registers and zero those in the panic data not
to leak the secrets.

BUG=b:320388543
TEST=Crash the FPMCU with the 'crash' console command and make sure the
general purpose registers are zeroed and are not logged on the console

Change-Id: I4b271d29b23c81a686a11e8db6023fa350d1d871
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5200957
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Dawid Niedźwiecki <dawidn@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Patryk Duda <patrykd@google.com>