flash_ec: Fix FMAP section detection
When reading the FMAP section out of the currently running EC image, the
script was using "dump_fmap" to verify it was valid.  However,
"dump_fmap" succeeds as long as there is an FMAP header anywhere in the
file, even if the extracted FMAP section is incomplete.

Switch to checking that FMAP header is found at the beginning of the
extracted file.

BUG=none
TEST=./util/flash_ec --brox --zephyr

Change-Id: I934b0612008ffe09f35bd90e5d8885a8b7dc74ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5269627
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>