pdc: Improve robustness of error handling for send_cmd
This CL modifies how the CCI_ERROR case is handled when waiting for a
send command operation to complete. In the case of an error, the PDC
may set both the error and complete bits of the PING_STATUS. The
send_cmd_wait_run state will check and clear the CCI_ERROR bit, but
the CCI_COMPLETE bit will remain set.

In addition, the CCI_ERROR bit causes the PDC command to be resent and
the PDC state remains in pdc_send_cmd_wait. When this SM is run again,
the CCI_COMPLETE bit is set from the previous error case, but will be
treated as the complete indication for the command that was resent.

This can lead to issues where the PDC subsytem gets out of sync with
items such as connector status and will remain in either the
SNK_ATTACHED or SRC_ATTACHED states even though there is no active
USB-C connection.

To fix this issue, this CL ensures that the error handling case in
send_cmd_wait clears the CCI_COMPLETED bit. Also, a new variable is
used to track how many times a particular command has been attempted
to be sent. The previous varialbe retry_counter is changed to
wait_counter as in practice it is increasing each time send_cmd_wait
is executed.

BUG=b:325070749
TEST=Verified that the PDC state remains in sync with the USB-C
connection. In addition, verified that commands which result in an
error are only resent a maximum number of 2 times
LOW_COVERAGE_REASON=Coverage to be handled in b/317065172

Change-Id: Ia7d0f90759fd2195790d9a91ef449e20377f59ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5293300
Commit-Queue: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Sam Hurst <shurst@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Eric Peers <epeers@google.com>