llvm: always enable TOOLCHAIN_HAS_NEWLIB
In the hardware model v2 the ARCH variable is not available at the point
where the toolchain config is evaluated. Instead, ARCH is set later from
the Kconfig, but in this case it is used to set the toolchain support
which is also going to change what's in Kconfig, creating a circular
dependency.

Just enable newlib support unconditionally add add a note that the
library may not actually be there for some architecture.

BUG=none
TEST=zmake compare-builds -a

Change-Id: Ic08a2c86048b4843e236b9aae0cff666013f4daf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5297714
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>