UCSI/PPM: Fix incorrect handling of async event
If an async event is sent while waiting for ack_cc_command, it will get
lost because it is not a pending command. In order to properly handle
async events that are sent mid-command cycle, we need to skip waiting on
the condvar if there are pending commands or events.

BUG=b:319304062
TEST=Unit test + validate with kernel driver

Change-Id: I3c684cd6e1ef826e64e5fb89ceb3d97e6d132ec6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5350394
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-by: Pavan Holla <pholla@google.com>
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>