board/hyperdebug: Handle cases of I2C interference
HyperDebug can be told to respond to a particular I2C address, which is
useful when testing GSC acting as I2C host on auxiliary busses.

Until now, temporary interference from e.g. other I2C devices responding
to the same address could cause HyperDebug to enter an interrupt loop,
as the existing code did not recognize the "arbitration loss" status
bit.  This CL make the HyperDebug firmware recognize the situation, and
record that there was interference on the bus, such that when
opentitantool later retrieves a transcript of what happened on the bus,
it gets a chance to react.  (And at the very least, HyperDebug will not
hit watchdog reset.)

BUG=b:319123634
TEST=make tast TAST_EXPR=gscdevboard.GSCI2CBridge.cap_always  (with LED driver chip also on I2C bus)

Change-Id: Idc3a44843fc512ff4fbd2000fe23adfcba98cb0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5347675
Tested-by: Jes Klinke <jbk@chromium.org>
Commit-Queue: Jes Klinke <jbk@chromium.org>
Reviewed-by: Jett Rink <jettrink@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>