chip/npcx: Clear SPR0 bit on invalid protected range
Fix situation when flash protect bit (SRP0) is enabled, but the size
of protected area is 0 or it's not possible to decode protected range
from SR1 and SR2 registers (spi_flash_reg_to_protect() returned
error). This situation can occur if flashing was interrupted
e.g. flashrom was killed while reading from flash.

Status registers can be modified only when the SRP0 bit and the WP_IF
bit (in DEV_CTL4 register) are not enabled at the same time. The
WP_IF bit is cleared when MCU reboots, it means that once enabled,
the bit can't be cleared by the software.

The WP_IF bit is set by flash_protect_int_flash() function based on
GPIO_WP status. In our case, the WP_IF bit is clear in RO (because we
are after reboot), but not in RW (because it will be set later in
this function).

Clearing the status registers before the WP_IF bit is enabled avoids
situation in which we protect status registers with size of protected
area set to 0. We rely on other parts of the system to enable
protection like we rely on them to enable protection when HW WP is
enabled for the first time.

BUG=b:328066864
TEST=Compile helipilot firmware and run it on DUT with helipilot board.
     Make sure the firmware works and there are no regressions.
TEST=Try to reproduce bug using steps from b/328066864#comment12
     Make sure the problem doesn't reproduce.

Change-Id: I21d52b5742e08d9f946541607d62285c3f043a3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5365866
Tested-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Patryk Duda <patrykd@google.com>
Reviewed-by: CH Lin <chlin56@nuvoton.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>