pdc/pd_task_intel_altmode: Correct mux config with inverted charger
Connecting a PD-only charger in a flipped orientation will currently
configure the mux in a connected, flipped orientation. This is incorrect
and interferes with S0ix. Ensure that the mux is disconnected and
configured to none no matter the orientation of the PD-only charger.

BUG=b:330105714
TEST=suspend_stress_test -c 10 in both orientations
LOW_COVERAGE_REASON=b:317065172

Change-Id: I4a2c9069d067d6fabea3e0e16209e08a3455af2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5410861
Tested-by: Robert Zieba <robertzieba@google.com>
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Commit-Queue: Robert Zieba <robertzieba@google.com>
Reviewed-by: Sam Hurst <shurst@google.com>
Auto-Submit: Robert Zieba <robertzieba@google.com>