third_party/boringssl: Don't define CROS_ZEPHYR and ANDROID
CROS_ZEPHYR is not a good define to indicate that we are building
BoringSSL in Zephyr environment. Instead, we should rely on __ZEPHYR__
defined by build system. Related BoringSSL change [1]

Also, we don't need to pretend that we are android - [2] and [3]

[1] https://boringssl-review.googlesource.com/c/boringssl/+/67627
[2] https://boringssl-review.googlesource.com/c/boringssl/+/65568
[3] https://boringssl-review.googlesource.com/c/boringssl/+/67667

BUG=b:321092852 b:273639386
TEST=CQ

Change-Id: I8c42380a3734be82b84cd203e19844e10967ed7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5527013
Tested-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Patryk Duda <patrykd@google.com>