zephyr/test/boringssl: Remove fatal error handler
CL:5173629 removed code that would cause a crash with k_oops and
replaced it with a getentropy() implementation which returns an
error in case of failure.

This patch removes fatal error handler which is a dead code now.

BUG=b:273639386
TEST=twister -v -i -T zephyr/test/ --test boringssl.selftest

Change-Id: I9f1e00e70343ca5832d1dd7ed9d747b814201709
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5483326
Tested-by: Patryk Duda <patrykd@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Patryk Duda <patrykd@google.com>