RTS54XX: Handle raw UCSI commands
With this CL, rts54_execute_ucsi_cmd will post a command as
CMD_RAW_UCSI. The driver skips command completion processes for
CMD_RAW_UCSI, which aren't needed for the PPM.

GET_CONNECTOR_STATUS needs to be handled differently because the out
data needs to be saved in a cache.

ACK_CC_CI is also customized for RTS54xx.

BUG=b:319124515
TEST=./twister -civ -T zephyr/test/pdc
LOW_COVERAGE_REASON=b:343188354

Change-Id: I368f952b21642c580841e3dbe76925d163fe19cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5512827
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>