board/hyperdebug: Declare pins on CN11 and CN12
Connectors CN11 and CN12 are by default unpopulated on the HyperDebug
board, but a few of the pins, which carries signals that are not also
present on any of the connectors already declared (CN7 - CN10), could be
used for custom circuitry, or when debugging test scripts using a logic
analyzer.

This CL adds declaratation of the above mentioned pins.

BUG=b:339034638
TEST=make BOARD=hyperdebug

Change-Id: Ifa969ca5d425bb1c28763f62fa0bfece0eee16aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5518113
Commit-Queue: Jes Klinke <jbk@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Jett Rink <jettrink@google.com>
Tested-by: Jes Klinke <jbk@chromium.org>