board/hyperdebug: Refactor interrupt vector table
The gpio monitoring feature relies ond hand-crafted machine code
handlers, which are updated in SRAM in a just-in-time fashion as the
monitoring is configured.

When not doing such monitoring, HyperDebug uses the "ordinary" EC RTOS
interrupt handling of GPIO edges, in order to react to the blue button
being pressed.

Until now, HyperDebug had a copy of the interrupt vector table in ram,
in which all the 16 entries relating to GPIO edges were permanently set
to point into SRAM, and managed the switch between the two "modes"
mentioned above by means of setting CORTEX_VTABLE to point either at the
SRAM copy, or at the default flash address.

This CL prepares for the bitbang feature being optimized to also use
hand-crafted machine code in SRAM for a timer interrupt handler, meaning
that we cannot set CORTEX_VTABLE to point at the flash table.

Instead this CL makes it such that CORTEX_VTABLE always points at the
SRAM copy, but that the 16 GPIO edge entries are switched back and forth
between the standard EC RTOS handler and the special SRAM routines.

BUG=b:341302705
TEST=make BOARD=hyperdebug  (manually run gpio monitoring)

Change-Id: I374487335b0b305f2c6effd9b49640b58113eecd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5565136
Reviewed-by: Jett Rink <jettrink@google.com>
Commit-Queue: Jes Klinke <jbk@chromium.org>
Tested-by: Jes Klinke <jbk@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>