charger: Modify PROCHOT threshold defaults
To prepare for moving the PROCHOT thresholds to the common driver code,
change the default setting to -1 to signal to the driver to use the
power on default setting and to not overwrite the PROCHOT thresholds.

BUG=b:344628812
TEST=zmake compare-builds -a
LOW_COVERAGE_REASON=Just a rename, no code changed.

Change-Id: I7ef3c1c071614cfd381528ca29c8ebd249520b55
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5593757
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Code-Coverage: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Reviewed-by: caveh jalali <caveh@chromium.org>