drivers: add zephyr/ prefix to autogenerated include
These files are now under a prefix. They can still be included on the
legacy path for now and upstream is not quite in the state where we can
flip the default, but this fixes the current paths for now.

BUG=none
TEST=cq dry run

Change-Id: I10a3fb59f739e267743b5058cee1f2c6ce79a0e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5604403
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>