pdc: ti: Add flag for entering driver suspend state
The idle run state blocks on a set of events leaving the driver suspend
request flag unchecked most of the time. Add an event to wake up the
thread while requesting the driver to suspend.

BUG=None
TEST=Run on Brox, PDC suspend works.
LOW_COVERAGE_REASON=Coverage to be handled in b/345292002

Change-Id: I0432cb5fc60ef04ac065ad391f051340e159ca77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5605643
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>