zephyr: add an option to use CrosEC sbrk function
Add an option to use the CrosEC version of the sbrk function.
It is used by default when newlib is present.

The CrosEC version, as opposite to the one in libc-hooks for newlib,
doesn't give memory at the end of RAM, where the jump data is stored.

BUG=b:324574796
TEST=make buildall && zmake build -a
LOW_COVERAGE_REASON=Tests added in CL:5639700 as a part of b:324574796

Change-Id: I674275962b6da006876cb06ee3542ce452306fd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5639699
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Patryk Duda <patrykd@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>