guybrush: Decrease max sensor frequency to 125Hz
The accel and gyro sensors are on the same i2c port as the thermal
sensors. When motion sense is running at 200hz the hook that reads
temperature sensors may deadlock waiting for the i2c port mutex and
cause a watchdog timeout. This only happens when other tasks are also
very active, such as the keyscan task. The actual max frequency may be
less depending on the step size of the sensor.

BUG=b:346492855
TEST=No watchdog with kpress + 100Hz motionsense + firmware update on
dewatt.
TEST=Pass cts modules CtsSensorTestCases and CtsAccelerationTestCases

Change-Id: I1e53c1d3eec09778aff089a73212f5115b342f68
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5646164
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Rob Barnes <robbarnes@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Rob Barnes <robbarnes@google.com>
Reviewed-by: Tim Van Patten <timvp@google.com>