keyboard_input: implement ksstate and kbpress
Implement some minimal ksstate and kbpress, there's upstream command for
the extra functionalities already but these specific ones are used in
faft and tast tests so let's implement a version compatible with the
legacy one.

Note that this is the third kbpress implementation so the one in mkbp
had to be filtered out as well.

BUG=b:274443029
TEST=build run on brox
TEST=cq dry run

Change-Id: I9e3825214d098ce7c5dec627d14ed575e80efbd5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5644968
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Yuval Peress <peress@google.com>