boot_keys, runtime_keys: make the config node a child of keyboard
Change the boot and runtime keys drivers so that they are instantiated
as a child node of the keyboard device, and use the child relashionship
rather than a chosen node. This makes the devicetree files less
fragemented and easier to follow.

BUG=none
TEST=zmake compare-builds brox greenbayupoc lotso
TEST=./twister -v -T zephyr/test/runtime_keys
TEST=./twister -v -T zephyr/test/boot_keys

Change-Id: I4213ad6b0c6a7ac018cb292758c17b1b3393c68f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5658535
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>