marigold: control CSIN SINK Discharge (#1057)
when plug-out ac should set 0x39 bit15 for fast discharge
until charger response acok level to low.

This is to solve the icon side effect issue when theAC is removed. Set charger 0x39 bit15 to "1" to speed up the sink input current. Because there is not set it back to "0", the charger continues to discharge in the AC state, increasing power consumption.

The solution proposed by the Power team is that the AC is unplugged, bit15 will be set to "1" first, and then set to "0" when the charger response acok level to low is reached, which can avoid this problem.

Signed-off-by: LeoCX_Tsai <LeoCX_Tsai@compal.com>
Co-authored-by: Daniel Schaefer <dhs@frame.work>