zephyr: ap_pwrseq: make the trigger struct a static const
The trigger structure is saved to the driver data and passed to the
callback. This is not currently used here but make it a static const
anyway so that we don't accidentally dereference a struct on the stack
in the future.

BUG=none
TEST=cq dry run

Change-Id: Ib8932bc68e9206820281b1fbee98fc178ce60988
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5667171
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Sam Hurst <shurst@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>