drivers: boot_keys: run the boot key check after lpc_init_mask()
The boot key check code works by calling host_set_single_event() to send
an event to the host. That call filters the events based on a mask
initialized by lpc_init_mask(), called in the app main(), so this code
needs to run after that mask is initialized but before vboot_main().

Change the init code from a SYS_INIT to a INIT_EARLY hook to ensure that
this runs in the right point in sequence.

BUG=b:350607363
TEST=enter recovery on brox

Change-Id: I6837ddd81c82d1fe74dc32fe2c9e639790ebca95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5673739
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: wen zhang <zhangwen6@huaqin.corp-partner.google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: wen zhang <zhangwen6@huaqin.corp-partner.google.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>