emul_tps6699x: Acount for length byte in messages
The first data byte of I2C write transactions to the TPS6699x is a
length byte. Do not treat this byte as data to be written, but record it
for later comparison to the number of bytes actually written.

BUG=b:303768388
TEST=twister -s pdc/pdc.generic.tps6699x; does not oops
LOW_COVERAGE_REASON=Coverage being added in b:345292002

Change-Id: I3a2bb2e805a48a127fd94159ddda122f79dca16d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5714617
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Abe Levkoy <alevkoy@chromium.org>
Reviewed-by: Sam Hurst <shurst@google.com>
Commit-Queue: Abe Levkoy <alevkoy@chromium.org>