fwk: lilac: add EC_FP_EN in GPIOA7
Add EC_FP_EN and following MODULE_PWR_EN to enable/disable.
follow the schematic net name to modify the GPIO node name.

Currently `gpio_ec_fp_en` is unconditionally enabled. In the future we
want to enable it as soon as possible after the power button is pressed
to improve user experience. We don't want to wait until the CPU starts up.

`tp_board_id_pug` will be used in the future to detect if the touchpad
has been plugged in, without having to turn on the 3V power to the
touchpad.

BRANCH=fwk-lilac-22648
TEST=zmake build azalea
BUG=None

Signed-off-by: johnwc_yeh <JohnWC_Yeh@compal.com>
Signed-off-by: Daniel Schaefer <dhs@frame.work>