common: Add variadic cprint functions
When integrating private libraries, we need to be able to define
proxy print/log functions that use the same print/log mechanisms.
These variadic functions all writing such proxy functions.
An example of these proxy log functions can be seen in
https://crrev.com/c/5725840.

The 0 argument to the printf attribute is documented for use with
vprint* functions:
https://gcc.gnu.org/onlinedocs/gcc-3.2/gcc/Function-Attributes.html
Clang's format for printf says to see GCC's documentation:
https://clang.llvm.org/docs/AttributeReference.html#format

We can't really test the debug_printf.c changes, since this code doesn't
seem to be used anywhere. The only mention of this code is in the
chocodile_vpdmcu board, but it is disabled by macro conditionals.
If you manually enable this debug_printf in that board config, the code
doesn't seem to compile even before this change.

BUG=b:343545905
TEST=make buildall; echo Result was $?
TEST=make proj-chocodile_vpdmcu
TEST=zmake build --clobber --all; echo Result was $?
TEST=Tested print functionaity on buccaneer.

Change-Id: I01485035602622183add8db8282c042a479163a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5735898
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: Craig Hesling <hesling@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>