Kconfig.defaults: enable I2C_ALLOW_NO_STOP_TRANSACTIONS by default
Some of the operations common/i2c_controller.c depends on the I2C APIs
being able to handle transactions with no stop at the end.

This is against the Zephyr I2C API definition as it breaks atomicitiy
between transactions and may result in the bus left in inconsistent
state, but at this stage the EC code depends on it.

The option allows to use the old behavior, this is a temporary stopgap
until the underlying API misuse is being reworked and fixed.

BUG=none
TEST=./twister

Cq-Depend: chromium:5757465
Cq-Depend: chromium:5882735
Change-Id: I3e5db436f5b0ebd0f07679f255416d0be1df64e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5763259
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>