fwk: implement the battery extender feature
The purpose of this function is to automatically drop the battery max
charge voltage if the system is left plugged in to AC for a long time
to preserve battery life.

When the timer is exceeded the battery charge voltage should be reduced
to 4.35V/cell. Two additional days after the timer has exceeded, the
battery charge voltage shall be reduced to 4.3V/cell.

The timer should reset to 0 days any time the system is in S0/S0ix and
not attached to AC for 30 minutes or longer.

BRANCH=marigold
BUG=https://app.clickup.com/t/86eq06zen
TEST=when timer expired, the charge voltage will reduce to 97% and
96% battery maximum charge voltage
TEST=charge voltage will recover when the user unplugs the adapter
for 30 minutes when the system in S0/S0ix.

Signed-off-by: Josh-Tsai <josh_tsai@compal.com>