servov4: servov4p1: Fix argv out-of-bounds bug in `cc` console command
The `cc` console command takes an optional 2nd argument for controlling
the USB-PD CC polarity, but it does not check `argc` before accessing
it. This causes an occasional Servo crash when `argv[2]` is NULL and
gets passed to `strcasecmp`. This in turn causes `servod` to crash or
malfunction and interrupt the rest of the TAST test flow.

BUG=b:361143401
TEST=BOARD=servo_v4 make
TEST=BOARD=servo_v4p1 make
TEST=Whole TAST-PD run passes

Change-Id: Ibc339f6d5332ccaa5dce344c2d139993d5f0fd55
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5805639
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>