fwk: sunflower: touchpad module power and keyscan control
the keyboard scan task read the matrix even if the keyboard module
does not connect or no power. It causes lots of I2C BUS error message
and then trigger the watchdog.

This change disable the keyboard scan task and enable it when EC turns
on the module power to avoid the watchdog.

BRANCH=sunflower
BUG=https://app.clickup.com/t/86eq852p9
TEST=keyboard function is workable with PR#1114 and PR#1115

Signed-off-by: Josh-Tsai <josh_tsai@compal.com>