fwk: sunflower: workaround: disable lid function to avoid the system hang
When the system receives the lid open event at S0ix state, the system cannot
wake up. Need to remove the battery and adapter to recover.

The root cause still not be found, add the workaround to avoid this issue.
Should be reverted until the issue is solved

BRANCH=sunflower
BUG=https://app.clickup.com/t/86eq8588c
TEST=check lid function is disabled on sunflower mainboard

Signed-off-by: Josh-Tsai <josh_tsai@compal.com>