trulo: ndorisa specific cbi support
This CL adds support for the EEPROM CBI which is on ndorisa, but is
not part of the Trulo reference design. In addition, it configures CBI
to use flash instead of the EEPROM.

BUG=b:357938662
TEST=zmake build trulo
Verifed that the CBI contents remained the same after disabling CBI
EEPROM config options.

Change-Id: Ifecdb639718b11fa56fec8bdba234016d15e1411
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5867310
Reviewed-by: caveh jalali <caveh@chromium.org>
Tested-by: Scott Collyer <scollyer@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Scott Collyer <scollyer@chromium.org>