pdc_tps6699x: Interrupt firing multiple times per IRQ
The PDC shares the same irq-gpio config, so during pdc_init,
the interrupt callback is added per pdc entry in device tree.
This leads it be be added twice.  When the IRQ occurs, this causes
the callback to be called twice.  This loop to notify all ports is
not needed.

BUG=None
TEST=./twister -c -T zephyr/test/pdc

Change-Id: I6dd9d55866787b3b623901b116d281b7c43c7a71
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5857937
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>