fwk: sunflower: workaround: shutdown the system before EC reboot
An unknown issue causes the EC to be stuck after the EC reboot.
We found this issue is related to the ec_rsmrst_l signal, it only happens
once the ec_rsmrst_l sets high. Still need the hardware team and EC vendor
help to clarify.

Add the workaround to force shutdown of the system to avoid EC being stuck.

BRANCH=fwk-sunflower-26784
BUG=None
TEST=system will auto reboot after executing EC_CMD_REBOOT_EC

Signed-off-by: Josh-Tsai <josh_tsai@compal.com>