test: Migrate fpupload command unit test to fpsensor_debug
The unit tests for the functions defined in fpsensor_debug.cc were added
in restricted_console unit test file. Migrate the unit test of the
fpupload, fpdownload, and fpmatch commands to the fpsensor_debug unit
test because this should be the dedicated file for all the tests related
to the functions defined in that source file.

BUG=b:369847148
TEST=./test/run_device_tests.py -b bloonchipper -t fpsensor_debug

Change-Id: I52906ecde80c1c1dacac0010c230a8ada2819cf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5892969
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Firas Sammoura <fsammoura@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Tested-by: Firas Sammoura <fsammoura@google.com>