keyboard: clean up some Kconfig options
The dep from PM_DEVICE_RUNTIME was incorrectly set on the col2 driver,
should have been on the whole keyboard one. Move it to the right
location, and also convert it to a "select" so it does not have to be
picked manually.

BUG=none
TEST=zmake compare-builds -a

Change-Id: I6b52aa4385e7adec965fe5bad032a5ec811fa1c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5907813
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: caveh jalali <caveh@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>