fwk: sunflower: Instant power button LED feeback
SLP_S5# signal is de-asserted only two seconds after the power button
was pressed and the CPU has started turning on. This signal is currently
used to determine the system state and also to turn on the power button
LED when the system turns on. On other platforms this signal is
immediately de-asserted, so the power button LED fedback is fine.
On sunflower it's not, so we need to manually enable it earlier.

The power button must illuminate immediately when EC registers the
power on event.

BUG=https://app.clickup.com/t/86eqcvh4p
BRANCH=fwk-sunflower-26784
TEST=press power button, the power led turns on immediately

Signed-off-by: Josh-Tsai <josh_tsai@compal.com>