cmake: toolchain: llvm: Provide LLVM_TOOLCHAIN_PATH
LLVM_TOOLCHAIN_PATH is used as a base path to look for 'newlib.cfg' or
'picolibc.cfg' provided by toolchain, which contains necessary options
to use non-default C library. If no config file is found, Zephyr will
assume that selected library is the default one.

In our case this is not actually useful because we have wrappers
(e.g. /usr/bin/armv7m-cros-eabi-clang) that run /usr/bin/clang with
appropriate options. The reason we define the variable is that CMake
looks for the file in / if the variable is empty, which causes the build
to be slow and cause problems if the file was found somewhere.

BUG=b:373472978
TEST='zmake build bloonchipper' is not slow

Change-Id: If9ad43c73a2a7cc77cc217798a7466d95c44b67e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5937941
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Patryk Duda <patrykd@google.com>
Tested-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>